Add Malayalam localization and Reports feature
- Add Malayalam script support for nakshatras across Devotee, Booking pages
- Add Malayalam month names to Panchang calendar (ചിങ്ങം, കന്നി, etc.)
- Create shared malayalam.js utility for consistent translations
- Add Reports page with income/expense summaries and date filtering
- Fix calendar month navigation bug (string concatenation issue)
- Update CLAUDE.md with Malayalam utility documentation

+++ app/Http/Controllers/Web/PanchangController.php
@@ -15,8 +15,8 @@ class PanchangController extends Controller
 {
     public function index(): Response
     {
-        $month = request('month', now()->month);
-        $year = request('year', now()->year);
+        $month = (int) request('month', now()->month);
+        $year = (int) request('year', now()->year);
 
         // Get panchang data for the selected month
         $startDate = Carbon::create($year, $month, 1)->startOfMonth();
